+++ app/src/main/java/com/example/githubuserfilter/userdetails/domain/model/DetailUserInfo.kt
@@ -1,13 +1,13 @@
 package com.example.githubuserfilter.userdetails.domain.model
 
 data class DetailUserInfo(
-    val id: Int,
+    val id: Long,
     val username: String,
     val name: String,
     val avatarUrl: String,
     val type: String,
     val email: String,
     val bio: String,
-    val followers: Int,
-    val publicRepos: Int
+    val followers: Long,
+    val publicRepos: Long
 )
